chg: renamed project and code cleanup
* renamed the project to 'Streaming SPARQL' to avoid name collusions
* implement the `Closable` interface for `StreamingResultSet`
* apply google code formatter
