--- src/main/java/ch/htwchur/isparql/console/package-info.java
@@ -1,8 +1,0 @@
-
-/**
- * Example iSparql command line clients.
- * 
- * @author albert.weichselbraun@htwchur.ch
- *
- */
-package ch.htwchur.isparql.console;
