--- src/main/java/ch/htwchur/isparql/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Incremental processing of SPARQL responses.
- * 
- * @author Albert Weichselbraun
- *
- */
-package ch.htwchur.isparql;
