+++ src/main/java/ch/htwchur/sparql/console/package-info.java
@@ -1,0 +1,7 @@
+
+/**
+ * Example iSparql command line clients.
+ *
+ * @author albert.weichselbraun@htwchur.ch
+ */
+package ch.htwchur.sparql.console;
