+++ src/main/java/ch/htwchur/sparql/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Incremental processing of SPARQL responses.
+ *
+ * @author Albert Weichselbraun
+ */
+package ch.htwchur.sparql;
